pyproject.toml
[tool.poetry]
name = "xraptor"
version = "0.2.0"
description = "Fast as websocket easy as http"
authors = ["Marco Sievers de Almeida <im.ximit@gmail.com>"]
readme = "README.md"
packages = [
    { include = "xraptor"},
]

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^12.0"
python-ulid = "^2.7.0"
witch-doctor = "^1.2.0"
python-decouple = "^3.8"
redis = { version = "^5.0.8", optional = true}
meeseeks-singleton = "^0.4.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
black = "^23.11.0"
pylint = "^2.15.10"
pre-commit = "^3.5.0"

[tool.poetry.extras]
redis_version = ["redis"]

[tool.poetry.urls]
homepage = "https://github.com/CenturyBoys/x-raptor"
repository = "https://github.com/CenturyBoys/x-raptor"
documentation = "https://github.com/CenturyBoys/x-raptor"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
